pyproject.toml
[tool.poetry]
name = "django_orm"
version = "0.1.0"
description = "Django ORM Implementation"
authors = ["NumanIbnMazid <numanibnmazid@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
autopep8 = "^1.6.0"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
